rename module to future_metrics
